soc on add-recipe
