+++ src/app/upload/components/add-recipe.tsx
@@ -4,14 +4,11 @@
  * `AddRecipeForm` is a React functional component that renders a form for adding new recipes.
  * It utilizes the `useFormState` hook to manage form state and handle submission.
  * Each field is created using the `InputForm` component.
- * A submit button is provided to submit the form data.
- *
- * The form uses Tailwind CSS classes for styling.
  */
 
-import { createRecipe } from '@/app/upload/action';
-import Button from '@/app/upload/components/button';
 import { useFormState } from 'react-dom';
+import { createRecipe } from '@/app/upload/actions/createRecipe';
+import Button from '@/app/upload/components/button';
 import InputForm from '@/app/upload/components/input-form';
 
 export default function AddRecipeForm() {
